+++ gantry/commands/run.py
@@ -78,10 +78,8 @@ from .main import CLICK_COMMAND_DEFAULTS, main
 @click.option(
     "--beaker-image",
     type=str,
-    default=constants.DEFAULT_IMAGE,
-    help="""The name or ID of an image on Beaker to use for your experiment.
-    Mutually exclusive with --docker-image.""",
-    show_default=True,
+    help=f"""The name or ID of an image on Beaker to use for your experiment.
+    Mutually exclusive with --docker-image. Defaults to '{constants.DEFAULT_IMAGE}' if neither is set.""",
 )
 @click.option(
     "--docker-image",
@@ -270,7 +268,7 @@ def run(
     workspace: Optional[str] = None,
     cluster: Optional[Tuple[str, ...]] = None,
     hostname: Optional[Tuple[str, ...]] = None,
-    beaker_image: Optional[str] = constants.DEFAULT_IMAGE,
+    beaker_image: Optional[str] = None,
     docker_image: Optional[str] = None,
     cpus: Optional[float] = None,
     gpus: Optional[int] = None,
@@ -317,7 +315,9 @@ def run(
             "[ARGS]... are required! For example:\n$ gantry run -- python -c 'print(\"Hello, World!\")'"
         )
 
-    if (beaker_image is None) == (docker_image is None):
+    if beaker_image is None and docker_image is None:
+        beaker_image = constants.DEFAULT_IMAGE
+    elif (beaker_image is None) == (docker_image is None):
         raise ConfigurationError(
             "Either --beaker-image or --docker-image must be specified, but not both."
         )